rename feature file
